+++ base/cvd/cuttlefish/host/commands/vhost_user_media/emulated_camera_mplane/src/device.rs
@@ -12,51 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::collections::VecDeque;
 use std::io::BufWriter;
-use std::io::Seek;
-use std::io::SeekFrom;
 use std::io::Write;
-use std::os::fd::AsFd;
-use std::os::fd::BorrowedFd;
 
 use v4l2r::PixelFormat;
 use v4l2r::QueueType;
 use v4l2r::bindings;
-use v4l2r::bindings::v4l2_fmtdesc;
 use v4l2r::bindings::v4l2_format;
-use v4l2r::bindings::v4l2_frmsizeenum;
-use v4l2r::bindings::v4l2_frmivalenum;
-use v4l2r::bindings::v4l2_requestbuffers;
-use v4l2r::ioctl::BufferCapabilities;
-use v4l2r::ioctl::BufferField;
-use v4l2r::ioctl::BufferFlags;
-use v4l2r::ioctl::V4l2Buffer;
-use v4l2r::memory::MemoryType;
-use virtio_media::VirtioMediaDevice;
-use virtio_media::VirtioMediaDeviceSession;
-use virtio_media::VirtioMediaEventQueue;
-use virtio_media::VirtioMediaHostMemoryMapper;
-use virtio_media::io::ReadFromDescriptorChain;
-use virtio_media::io::WriteToDescriptorChain;
+use vhu_media::devices::CaptureDeviceFormat;
 use virtio_media::ioctl::IoctlResult;
-use virtio_media::ioctl::VirtioMediaIoctlHandler;
-use virtio_media::ioctl::virtio_media_dispatch_ioctl;
-use virtio_media::memfd::MemFdBuffer;
-use virtio_media::mmap::MmapMappingManager;
-use virtio_media::protocol::DequeueBufferEvent;
-use virtio_media::protocol::SgEntry;
-use virtio_media::protocol::V4l2Event;
-use virtio_media::protocol::V4l2Ioctl;
-use virtio_media::protocol::VIRTIO_MEDIA_MMAP_FLAG_RW;
-
-use vhu_media::devices::{
-    Buffer, BufferState, set_plane_offset_and_length,
-};
 
 const WIDTH: u32 = 640;
 const HEIGHT: u32 = 480;
 const FRAME_RATE: u32 = 30;
+const PIXELFORMAT: u32 = PixelFormat::from_fourcc(b"YU12").to_u32();
+const BUFFER_SIZE: u32 = WIDTH * HEIGHT * 3 / 2;
 
 const INPUTS: [bindings::v4l2_input; 1] = [bindings::v4l2_input {
     index: 0,
@@ -65,28 +35,56 @@ const INPUTS: [bindings::v4l2_input; 1] = [bindings::v4l2_input {
     ..unsafe { std::mem::zeroed() }
 }];
 
+pub struct MplaneFormat;
 
-/// Session data of [`EmulatedCamera`].
-pub struct EmulatedCameraSession {
-    /// Id of the session.
-    id: u32,
-    /// Current iteration of the pattern generation cycle.
-    iteration: u64,
-    /// Buffers currently allocated for this session.
-    buffers: Vec<Buffer>,
-    /// Queue of buffers awaiting processing.
-    queued_buffers: VecDeque<usize>,
-    /// Is the session currently streaming?
-    streaming: bool,
-}
+impl CaptureDeviceFormat for MplaneFormat {
+    const QUEUE_TYPE: QueueType = QueueType::VideoCaptureMplane;
+    const PIXELFORMAT: u32 = PIXELFORMAT;
+    const BUFFER_SIZE: u32 = BUFFER_SIZE;
+    const FRAME_RATE: u32 = FRAME_RATE;
+    const WIDTH: u32 = WIDTH;
+    const HEIGHT: u32 = HEIGHT;
+    const INPUTS: &'static [bindings::v4l2_input] = &INPUTS;
 
-impl VirtioMediaDeviceSession for EmulatedCameraSession {
-    fn poll_fd(&self) -> Option<BorrowedFd<'_>> {
-        None
+    fn default_fmt(queue: QueueType) -> v4l2_format {
+        let pix_mp = bindings::v4l2_pix_format_mplane {
+            width: WIDTH,
+            height: HEIGHT,
+            pixelformat: PIXELFORMAT,
+            field: bindings::v4l2_field_V4L2_FIELD_NONE,
+            colorspace: bindings::v4l2_colorspace_V4L2_COLORSPACE_SRGB,
+            num_planes: 3,
+            plane_fmt: [
+                bindings::v4l2_plane_pix_format {
+                    sizeimage: WIDTH * HEIGHT,
+                    bytesperline: WIDTH,
+                    ..Default::default()
+                },
+                bindings::v4l2_plane_pix_format {
+                    sizeimage: WIDTH * HEIGHT / 4,
+                    bytesperline: WIDTH / 2,
+                    ..Default::default()
+                },
+                bindings::v4l2_plane_pix_format {
+                    sizeimage: WIDTH * HEIGHT / 4,
+                    bytesperline: WIDTH / 2,
+                    ..Default::default()
+                },
+                Default::default(),
+                Default::default(),
+                Default::default(),
+                Default::default(),
+                Default::default(),
+            ],
+            ..Default::default()
+        };
+
+        v4l2_format {
+            type_: queue as u32,
+            fmt: bindings::v4l2_format__bindgen_ty_1 { pix_mp },
+        }
     }
-}
 
-impl EmulatedCameraSession {
     fn write_pattern<W: std::io::Write>(iteration: u64, sink: W) -> IoctlResult<()> {
         let mut writer = BufWriter::new(sink);
         let y = (iteration % 256) as u8;
@@ -104,511 +102,4 @@ impl EmulatedCameraSession {
         Ok(())
     }
 
-    /// Write basic pattern into the queued buffers
-    fn process_queued_buffers<Q: VirtioMediaEventQueue>(
-        &mut self,
-        evt_queue: &mut Q,
-    ) -> IoctlResult<()> {
-        while let Some(buf_id) = self.queued_buffers.pop_front() {
-            let iteration = self.iteration;
-            let buffer = self.buffers.get_mut(buf_id).ok_or(libc::EIO)?;
-            buffer
-                .fd
-                .as_file()
-                .seek(SeekFrom::Start(0))
-                .map_err(|_| libc::EIO)?;
-
-            Self::write_pattern(iteration, buffer.fd.as_file())?;
-
-            buffer.set_state(
-                BufferState::Outgoing {
-                    sequence: iteration as u32,
-                },
-                BUFFER_SIZE,
-            );
-            evt_queue.send_event(V4l2Event::DequeueBuffer(DequeueBufferEvent::new(
-                self.id,
-                buffer.v4l2_buffer.clone(),
-            )));
-
-            self.iteration += 1;
-        }
-
-        Ok(())
-    }
-}
-
-/// Emulated camera used for testing Android camera stack.
-///
-/// This implementation looks forward to have feature parity with existing Android Guest Emulated
-/// camera https://cs.android.com/android/platform/superproject/main/+/main:hardware/google/camera/devices/EmulatedCamera/
-pub struct EmulatedCamera<Q: VirtioMediaEventQueue, HM: VirtioMediaHostMemoryMapper> {
-    /// Queue used to send events to the guest.
-    evt_queue: Q,
-    /// Host MMAP mapping manager.
-    mmap_manager: MmapMappingManager<HM>,
-    /// ID of the session with allocated buffers, if any.
-    ///
-    /// v4l2-compliance checks that only a single session can have allocated buffers at a given
-    /// time, since that's how actual hardware works - no two sessions can access a camera at the
-    /// same time. It will fails if we allow simultaneous sessions to be active, so we need this
-    /// artificial limitation to make it pass fully.
-    active_session: Option<u32>,
-}
-
-impl<Q, HM> EmulatedCamera<Q, HM>
-where
-    Q: VirtioMediaEventQueue,
-    HM: VirtioMediaHostMemoryMapper,
-{
-    pub fn new(evt_queue: Q, mapper: HM) -> Self {
-        Self {
-            evt_queue,
-            mmap_manager: MmapMappingManager::from(mapper),
-            active_session: None,
-        }
-    }
-}
-
-impl<Q, HM, Reader, Writer> VirtioMediaDevice<Reader, Writer> for EmulatedCamera<Q, HM>
-where
-    Q: VirtioMediaEventQueue,
-    HM: VirtioMediaHostMemoryMapper,
-    Reader: ReadFromDescriptorChain,
-    Writer: WriteToDescriptorChain,
-{
-    type Session = EmulatedCameraSession;
-
-    fn new_session(&mut self, session_id: u32) -> std::result::Result<Self::Session, i32> {
-        Ok(EmulatedCameraSession {
-            id: session_id,
-            iteration: 0,
-            buffers: Default::default(),
-            queued_buffers: Default::default(),
-            streaming: false,
-        })
-    }
-
-    fn close_session(&mut self, session: Self::Session) {
-        // Nothing to cleanup when `close_session` is called for sessions without
-        // allocated buffers, hence the early return.
-        if self.active_session != Some(session.id) {
-            return;
-        }
-
-        self.active_session = None;
-
-        for buffer in &session.buffers {
-            self.mmap_manager.unregister_buffer(buffer.offset);
-        }
-    }
-
-    fn do_ioctl(
-        &mut self,
-        session: &mut Self::Session,
-        ioctl: V4l2Ioctl,
-        reader: &mut Reader,
-        writer: &mut Writer,
-    ) -> std::io::Result<()> {
-        virtio_media_dispatch_ioctl(self, session, ioctl, reader, writer)
-    }
-
-    fn do_mmap(
-        &mut self,
-        session: &mut Self::Session,
-        flags: u32,
-        offset: u32,
-    ) -> std::result::Result<(u64, u64), i32> {
-        let buffer = session
-            .buffers
-            .iter_mut()
-            .find(|b| b.offset == offset)
-            .ok_or(libc::EINVAL)?;
-        let rw = (flags & VIRTIO_MEDIA_MMAP_FLAG_RW) != 0;
-        let fd = buffer.fd.as_file().as_fd();
-        let (guest_addr, size) = self
-            .mmap_manager
-            .create_mapping(offset, fd, rw)
-            .map_err(|_| libc::EINVAL)?;
-        Ok((guest_addr, size))
-    }
-
-    fn do_munmap(&mut self, guest_addr: u64) -> std::result::Result<(), i32> {
-        self.mmap_manager
-            .remove_mapping(guest_addr)
-            .map(|_| ())
-            .map_err(|_| libc::EINVAL)
-    }
-}
-
-const PIXELFORMAT: u32 = PixelFormat::from_fourcc(b"YU12").to_u32();
-const BUFFER_SIZE: u32 = WIDTH * HEIGHT * 3 / 2;
-
-fn default_fmt(queue: QueueType) -> v4l2_format {
-    let pix_mp = bindings::v4l2_pix_format_mplane {
-        width: WIDTH,
-        height: HEIGHT,
-        pixelformat: PIXELFORMAT,
-        field: bindings::v4l2_field_V4L2_FIELD_NONE,
-        colorspace: bindings::v4l2_colorspace_V4L2_COLORSPACE_SRGB,
-        num_planes: 3,
-        plane_fmt: [
-            bindings::v4l2_plane_pix_format {
-                sizeimage: WIDTH * HEIGHT,
-                bytesperline: WIDTH,
-                ..Default::default()
-            },
-            bindings::v4l2_plane_pix_format {
-                sizeimage: WIDTH * HEIGHT / 4,
-                bytesperline: WIDTH / 2,
-                ..Default::default()
-            },
-            bindings::v4l2_plane_pix_format {
-                sizeimage: WIDTH * HEIGHT / 4,
-                bytesperline: WIDTH / 2,
-                ..Default::default()
-            },
-            Default::default(),
-            Default::default(),
-            Default::default(),
-            Default::default(),
-            Default::default(),
-        ],
-        ..Default::default()
-    };
-
-    v4l2_format {
-        type_: queue as u32,
-        fmt: bindings::v4l2_format__bindgen_ty_1 { pix_mp },
-    }
-}
-
-/// Implementations of the ioctls required by a v4l2 CAPTURE device.
-impl<Q, HM> VirtioMediaIoctlHandler for EmulatedCamera<Q, HM>
-where
-    Q: VirtioMediaEventQueue,
-    HM: VirtioMediaHostMemoryMapper,
-{
-    type Session = EmulatedCameraSession;
-
-    fn enum_fmt(
-        &mut self,
-        _session: &Self::Session,
-        queue: QueueType,
-        index: u32,
-    ) -> IoctlResult<v4l2_fmtdesc> {
-        if queue != QueueType::VideoCaptureMplane {
-            return Err(libc::EINVAL);
-        }
-        if index > 0 {
-            return Err(libc::EINVAL);
-        }
-
-        let fmtdesc = v4l2_fmtdesc {
-            index: 0,
-            type_: queue as u32,
-            pixelformat: PIXELFORMAT,
-            ..Default::default()
-        };
-
-        Ok(fmtdesc)
-    }
-
-    fn g_fmt(&mut self, _session: &Self::Session, queue: QueueType) -> IoctlResult<v4l2_format> {
-        if queue != QueueType::VideoCaptureMplane {
-            return Err(libc::EINVAL);
-        }
-        Ok(default_fmt(queue))
-    }
-
-    fn s_fmt(
-        &mut self,
-        _session: &mut Self::Session,
-        queue: QueueType,
-        _format: v4l2_format,
-    ) -> IoctlResult<v4l2_format> {
-        if queue != QueueType::VideoCaptureMplane {
-            return Err(libc::EINVAL);
-        }
-        Ok(default_fmt(queue))
-    }
-
-    fn try_fmt(
-        &mut self,
-        _session: &Self::Session,
-        queue: QueueType,
-        _format: v4l2_format,
-    ) -> IoctlResult<v4l2_format> {
-        if queue != QueueType::VideoCaptureMplane {
-            return Err(libc::EINVAL);
-        }
-        Ok(default_fmt(queue))
-    }
-
-    fn g_parm(
-        &mut self,
-        _session: &Self::Session,
-        queue: QueueType,
-    ) -> IoctlResult<bindings::v4l2_streamparm> {
-        if queue != QueueType::VideoCaptureMplane {
-            return Err(libc::EINVAL);
-        }
-
-        let mut parm = bindings::v4l2_streamparm {
-            type_: queue as u32,
-            ..Default::default()
-        };
-
-        // SAFETY: The `parm` union is used for the capture type.
-        let capture = unsafe { &mut parm.parm.capture };
-        capture.capability = bindings::V4L2_CAP_TIMEPERFRAME;
-        capture.timeperframe = bindings::v4l2_fract {
-            numerator: 1,
-            denominator: FRAME_RATE,
-        };
-
-        Ok(parm)
-    }
-
-    fn s_parm(
-        &mut self,
-        _session: &mut Self::Session,
-        mut parm: bindings::v4l2_streamparm,
-    ) -> IoctlResult<bindings::v4l2_streamparm> {
-        if parm.type_ != QueueType::VideoCaptureMplane as u32 {
-            return Err(libc::EINVAL);
-        }
-
-        // We just return the fixed values, ignoring what the user set.
-        // SAFETY: The `parm` union is used for the capture type.
-        let capture = unsafe { &mut parm.parm.capture };
-        capture.capability = bindings::V4L2_CAP_TIMEPERFRAME;
-        capture.timeperframe = bindings::v4l2_fract {
-            numerator: 1,
-            denominator: FRAME_RATE,
-        };
-
-        Ok(parm)
-    }
-
-    fn reqbufs(
-        &mut self,
-        session: &mut Self::Session,
-        queue: QueueType,
-        memory: MemoryType,
-        count: u32,
-    ) -> IoctlResult<v4l2_requestbuffers> {
-        if queue != QueueType::VideoCaptureMplane {
-            return Err(libc::EINVAL);
-        }
-        if memory != MemoryType::Mmap {
-            return Err(libc::EINVAL);
-        }
-        if session.streaming {
-            return Err(libc::EBUSY);
-        }
-        // Buffers cannot be requested on a session if there is already another session with
-        // allocated buffers.
-        match self.active_session {
-            Some(id) if id != session.id => return Err(libc::EBUSY),
-            _ => (),
-        }
-
-        // Reqbufs(0) is an implicit streamoff.
-        if count == 0 {
-            self.active_session = None;
-            self.streamoff(session, queue)?;
-        } else {
-            // TODO factorize with streamoff.
-            session.queued_buffers.clear();
-            for buffer in session.buffers.iter_mut() {
-                buffer.set_state(BufferState::New, BUFFER_SIZE);
-            }
-            self.active_session = Some(session.id);
-        }
-
-        let count = std::cmp::min(count, 32);
-
-        for buffer in &session.buffers {
-            self.mmap_manager.unregister_buffer(buffer.offset);
-        }
-
-        session.buffers = (0..count)
-            .map(|i| {
-                MemFdBuffer::new(BUFFER_SIZE as u64)
-                    .map_err(|e| {
-                        log::error!("failed to allocate MMAP buffers: {:#}", e);
-                        libc::ENOMEM
-                    })
-                    .and_then(|fd| {
-                        let offset = self
-                            .mmap_manager
-                            .register_buffer(None, BUFFER_SIZE)
-                            .map_err(|_| libc::EINVAL)?;
-
-                        let mut v4l2_buffer = V4l2Buffer::new(queue, i, MemoryType::Mmap);
-                        set_plane_offset_and_length(&mut v4l2_buffer, offset, BUFFER_SIZE);
-                        v4l2_buffer.set_field(BufferField::None);
-                        v4l2_buffer.set_flags(BufferFlags::TIMESTAMP_MONOTONIC);
-
-                        Ok(Buffer::new(v4l2_buffer, fd, offset))
-                    })
-            })
-            .collect::<std::result::Result<_, _>>()?;
-
-        Ok(v4l2_requestbuffers {
-            count,
-            type_: queue as u32,
-            memory: memory as u32,
-            capabilities: (BufferCapabilities::SUPPORTS_MMAP
-                | BufferCapabilities::SUPPORTS_ORPHANED_BUFS)
-                .bits(),
-            // This device does not support V4L2_BUF_CAP_SUPPORTS_MMAP_CACHE_HINTS,
-            // so the flags field in v4l2_requestbuffers must be 0.
-            flags: 0,
-            ..Default::default()
-        })
-    }
-
-    fn querybuf(
-        &mut self,
-        session: &Self::Session,
-        queue: QueueType,
-        index: u32,
-    ) -> IoctlResult<v4l2r::ioctl::V4l2Buffer> {
-        if queue != QueueType::VideoCaptureMplane {
-            return Err(libc::EINVAL);
-        }
-        let buffer = session.buffers.get(index as usize).ok_or(libc::EINVAL)?;
-
-        Ok(buffer.v4l2_buffer.clone())
-    }
-
-    fn qbuf(
-        &mut self,
-        session: &mut Self::Session,
-        buffer: v4l2r::ioctl::V4l2Buffer,
-        _guest_regions: Vec<Vec<SgEntry>>,
-    ) -> IoctlResult<v4l2r::ioctl::V4l2Buffer> {
-        let host_buffer = session
-            .buffers
-            .get_mut(buffer.index() as usize)
-            .ok_or(libc::EINVAL)?;
-        // Attempt to queue already queued buffer.
-        if matches!(host_buffer.state, BufferState::Incoming) {
-            return Err(libc::EINVAL);
-        }
-
-        host_buffer.set_state(BufferState::Incoming, BUFFER_SIZE);
-        session.queued_buffers.push_back(buffer.index() as usize);
-
-        let buffer = host_buffer.v4l2_buffer.clone();
-
-        if session.streaming {
-            session.process_queued_buffers(&mut self.evt_queue)?;
-        }
-
-        Ok(buffer)
-    }
-
-    fn streamon(&mut self, session: &mut Self::Session, queue: QueueType) -> IoctlResult<()> {
-        if queue != QueueType::VideoCaptureMplane || session.buffers.is_empty() {
-            return Err(libc::EINVAL);
-        }
-        session.streaming = true;
-
-        session.process_queued_buffers(&mut self.evt_queue)?;
-
-        Ok(())
-    }
-
-    fn streamoff(&mut self, session: &mut Self::Session, queue: QueueType) -> IoctlResult<()> {
-        if queue != QueueType::VideoCaptureMplane {
-            return Err(libc::EINVAL);
-        }
-        session.streaming = false;
-        session.queued_buffers.clear();
-        for buffer in session.buffers.iter_mut() {
-            buffer.set_state(BufferState::New, BUFFER_SIZE);
-        }
-
-        Ok(())
-    }
-
-    fn g_input(&mut self, _session: &Self::Session) -> IoctlResult<i32> {
-        Ok(0)
-    }
-
-    fn s_input(&mut self, _session: &mut Self::Session, input: i32) -> IoctlResult<i32> {
-        if input != 0 { Err(libc::EINVAL) } else { Ok(0) }
-    }
-
-    fn enuminput(
-        &mut self,
-        _session: &Self::Session,
-        index: u32,
-    ) -> IoctlResult<bindings::v4l2_input> {
-        INPUTS.get(index as usize).map(|&x| x).ok_or(libc::EINVAL)
-    }
-
-    fn enum_framesizes(
-        &mut self,
-        _session: &Self::Session,
-        index: u32,
-        pixel_format: u32,
-    ) -> IoctlResult<bindings::v4l2_frmsizeenum> {
-        if pixel_format != PIXELFORMAT {
-            return Err(libc::EINVAL);
-        }
-        if index > 0 {
-            return Err(libc::EINVAL);
-        }
-
-        Ok(v4l2_frmsizeenum {
-            index,
-            pixel_format,
-            type_: bindings::v4l2_frmsizetypes_V4L2_FRMSIZE_TYPE_DISCRETE,
-            __bindgen_anon_1: bindings::v4l2_frmsizeenum__bindgen_ty_1 {
-                discrete: bindings::v4l2_frmsize_discrete {
-                    width: WIDTH,
-                    height: HEIGHT,
-                },
-            },
-            ..Default::default()
-        })
-    }
-
-    fn enum_frameintervals(
-        &mut self,
-        _session: &Self::Session,
-        index: u32,
-        pixel_format: u32,
-        width: u32,
-        height: u32,
-    ) -> IoctlResult<bindings::v4l2_frmivalenum> {
-        if pixel_format != PIXELFORMAT {
-            return Err(libc::EINVAL);
-        }
-        if width != WIDTH || height != HEIGHT {
-            return Err(libc::EINVAL);
-        }
-        if index > 0 {
-            return Err(libc::EINVAL);
-        }
-
-        Ok(v4l2_frmivalenum {
-            index,
-            pixel_format,
-            width,
-            height,
-            type_: bindings::v4l2_frmivaltypes_V4L2_FRMIVAL_TYPE_DISCRETE,
-            __bindgen_anon_1: bindings::v4l2_frmivalenum__bindgen_ty_1 {
-                discrete: bindings::v4l2_fract {
-                    numerator: 1,
-                    denominator: FRAME_RATE,
-                },
-            },
-            ..Default::default()
-        })
-    }
 }
