+++ base/cvd/cuttlefish/host/commands/vhost_user_media/emulated_camera_mplane/src/main.rs
@@ -87,7 +87,12 @@ fn start_backend(config: Config) -> Result<()> {
         };
         let backend = Arc::new(RwLock::new(VhuMediaBackend::new(
             config,
-            |event_queue, host_mapper| crate::device::EmulatedCamera::new(event_queue, host_mapper),
+            |event_queue, host_mapper| {
+                vhu_media::devices::EmulatedCamera::<_, _, crate::device::MplaneFormat>::new(
+                    event_queue,
+                    host_mapper,
+                )
+            },
         )));
         let mut daemon = VhostUserDaemon::new(
             String::from("vhost-user-media-backend"),
